rename rules and stabilize tests
