+++ crates/ruff_linter/resources/test/fixtures/flake8_async/ASYNC110.py
