+++ crates/ruff_linter/resources/test/fixtures/flake8_async/ASYNC210.py
@@ -1,0 +1,23 @@
+import urllib.request
+import requests
+import httpx
+
+
+async def foo():
+    urllib.request.urlopen("http://example.com/foo/bar").read()
+
+
+async def foo():
+    requests.get()
+
+
+async def foo():
+    httpx.get()
+
+
+async def foo():
+    requests.post()
+
+
+async def foo():
+    httpx.post()
