--- crates/ruff_linter/resources/test/fixtures/flake8_trio/TRIO100.py
@@ -1,27 +1,0 @@
-import trio
-
-
-async def func():
-    with trio.fail_after():
-        ...
-
-
-async def func():
-    with trio.fail_at():
-        await ...
-
-
-async def func():
-    with trio.move_on_after():
-        ...
-
-
-async def func():
-    with trio.move_at():
-        await ...
-
-
-async def func():
-    with trio.move_at():
-        async with trio.open_nursery() as nursery:
-            ...
