+++ crates/ruff_linter/src/rules/flake8_trio/mod.rs
@@ -14,15 +14,15 @@ mod tests {
     use crate::settings::LinterSettings;
     use crate::test::test_path;
 
-    #[test_case(Rule::TrioTimeoutWithoutAwait, Path::new("TRIO100.py"))]
-    #[test_case(Rule::TrioSyncCall, Path::new("TRIO105.py"))]
-    #[test_case(Rule::TrioAsyncFunctionWithTimeout, Path::new("TRIO109.py"))]
-    #[test_case(Rule::TrioUnneededSleep, Path::new("TRIO110.py"))]
-    #[test_case(Rule::TrioZeroSleepCall, Path::new("TRIO115.py"))]
+    #[test_case(Rule::TrioTimeoutWithoutAwait, Path::new("ASYNC100.py"))]
+    #[test_case(Rule::TrioSyncCall, Path::new("ASYNC105.py"))]
+    #[test_case(Rule::TrioAsyncFunctionWithTimeout, Path::new("ASYNC109.py"))]
+    #[test_case(Rule::TrioUnneededSleep, Path::new("ASYNC110.py"))]
+    #[test_case(Rule::TrioZeroSleepCall, Path::new("ASYNC115.py"))]
     fn rules(rule_code: Rule, path: &Path) -> Result<()> {
         let snapshot = format!("{}_{}", rule_code.noqa_code(), path.to_string_lossy());
         let diagnostics = test_path(
-            Path::new("flake8_trio").join(path).as_path(),
+            Path::new("flake8_async").join(path).as_path(),
             &LinterSettings::for_rule(rule_code),
         )?;
         assert_messages!(snapshot, diagnostics);
