+++ ruff.schema.json
@@ -2693,10 +2693,18 @@
         "ASYNC1",
         "ASYNC10",
         "ASYNC100",
-        "ASYNC101",
-        "ASYNC102",
+        "ASYNC105",
+        "ASYNC109",
         "ASYNC11",
+        "ASYNC110",
+        "ASYNC115",
         "ASYNC116",
+        "ASYNC2",
+        "ASYNC21",
+        "ASYNC210",
+        "ASYNC22",
+        "ASYNC220",
+        "ASYNC222",
         "B",
         "B0",
         "B00",
@@ -3837,15 +3845,6 @@
         "TID251",
         "TID252",
         "TID253",
-        "TRIO",
-        "TRIO1",
-        "TRIO10",
-        "TRIO100",
-        "TRIO105",
-        "TRIO109",
-        "TRIO11",
-        "TRIO110",
-        "TRIO115",
         "TRY",
         "TRY0",
         "TRY00",
